+++ src/core/CMakeLists.txt
@@ -1,8 +1,7 @@
 
 cmake_minimum_required(VERSION 3.12)
 set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_FLAGS "-O3 -fPIC")
-add_compile_options("-std=gnu++17")
+set(CMAKE_CXX_FLAGS "-O3 -Wall -Wextra -funroll-loops")
 add_compile_options("-Wno-error=deprecated-declarations")
 link_libraries(stdc++fs)
