+++ src/core/include/xmlparser.hpp
@@ -36,13 +36,9 @@ static bool CONTAINS = false;
 
 void reinit(svector &);
 
-struct IDMap{
-    std::string name;
-    std::string node;
-    std::string key;
-    std::string degree;
-    svector values;
-};
+
+
+
 class XMLParser{
 
     private:
